# docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "SentinelZone AI/SOAR",
    "description": "Standalone investigation and human-approved SOAR platform. Dry-run results never represent live containment.",
    "version": "0.30.0-rc.1"
  },
  "paths": {
...
    "/v1/ai/analyze": {
      "post": {
        "tags": [
          "AI"
        ],
        "summary": "Analyze",
        "operationId": "analyze_v1_ai_analyze_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AnalyzeIn"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AIRunView"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "409": {
            "description": "Conflict",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "413": {
            "description": "Request Entity Too Large",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "422": {
            "description": "Unprocessable Entity",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "Internal Server Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "503": {
            "description": "Service Unavailable",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/v1/ai/runs/{run_id}": {
      "get": {
        "tags": [
          "AI"
        ],
        "summary": "Get Run",
        "operationId": "get_run_v1_ai_runs__run_id__get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "run_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Run Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AIRunView"
                }
              }
            }
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Unauthorized"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Forbidden"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Not Found"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Conflict"
          },
          "413": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Request Entity Too Large"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Unprocessable Entity"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Internal Server Error"
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Service Unavailable"
          }
        }
      }
    },
    "/v1/incidents/{incident_id}/analyses": {
      "get": {
        "tags": [
          "AI"
        ],
        "summary": "Analyses",
        "operationId": "analyses_v1_incidents__incident_id__analyses_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "incident_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Incident Id"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 200,
              "minimum": 1,
              "default": 50,
              "title": "Limit"
            }
          },
          {
            "name": "offset",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 0,
              "default": 0,
              "title": "Offset"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AIRunPage"
                }
              }
            }
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Unauthorized"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Forbidden"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Not Found"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Conflict"
          },
          "413": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Request Entity Too Large"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Unprocessable Entity"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Internal Server Error"
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Service Unavailable"
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "AIOutput": {
        "properties": {
          "observed_facts": {
            "items": {
              "$ref": "#/components/schemas/Fact"
            },
            "type": "array",
            "maxItems": 50,
            "title": "Observed Facts"
          },
          "hypotheses": {
            "items": {
              "$ref": "#/components/schemas/Hypothesis"
            },
            "type": "array",
            "maxItems": 20,
            "title": "Hypotheses"
          },
          "missing_evidence": {
            "items": {
              "type": "string",
              "maxLength": 2000,
              "minLength": 1
            },
            "type": "array",
            "maxItems": 100,
            "title": "Missing Evidence"
          },
          "recommended_investigation": {
            "items": {
              "type": "string",
              "maxLength": 2000,
              "minLength": 1
            },
            "type": "array",
            "maxItems": 30,
            "title": "Recommended Investigation"
          },
          "containment_options": {
            "items": {
              "type": "string",
              "maxLength": 2000,
              "minLength": 1
            },
            "type": "array",
            "maxItems": 20,
            "title": "Containment Options"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "observed_facts",
          "hypotheses",
          "missing_evidence",
          "recommended_investigation",
          "containment_options"
        ],
        "title": "AIOutput"
      },
      "AIRunPage": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/AIRunView"
            },
            "type": "array",
            "title": "Items"
          },
          "next_offset": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Next Offset"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "items",
          "next_offset"
        ],
        "title": "AIRunPage"
      },
      "AIRunView": {
        "properties": {
          "run_id": {
            "type": "string",
            "title": "Run Id"
          },
          "incident_id": {
            "type": "string",
            "title": "Incident Id"
          },
          "tenant_id": {
            "type": "string",
            "title": "Tenant Id"
          },
          "status": {
            "type": "string",
            "enum": [
              "running",
              "completed",
              "invalid",
              "rejected",
              "failed"
            ],
            "title": "Status"
          },
          "trusted": {
            "type": "boolean",
            "title": "Trusted"
          },
          "validation_status": {
            "type": "string",
            "enum": [
              "pending",
              "valid",
              "invalid",
              "unavailable"
            ],
            "title": "Validation Status"
          },
          "provider": {
            "type": "string",
            "title": "Provider"
          },
          "model": {
            "type": "string",
            "title": "Model"
          },
          "error_code": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Error Code"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "started_at": {
            "type": "string",
            "format": "date-time",
            "title": "Started At"
          },
          "completed_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Completed At"
          },
          "data_mode": {
            "type": "string",
            "enum": [
              "REAL",
              "TEST",
              "REPLAY",
              "UNKNOWN"
            ],
            "title": "Data Mode"
          },
          "observed_facts": {
            "anyOf": [
              {
                "items": {
                  "$ref": "#/components/schemas/Fact"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Observed Facts"
          },
          "hypotheses": {
            "anyOf": [
              {
                "items": {
                  "$ref": "#/components/schemas/Hypothesis"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Hypotheses"
          },
          "missing_evidence": {
            "anyOf": [
              {
                "items": {
                  "type": "string"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Missing Evidence"
          },
          "recommended_investigation": {
            "anyOf": [
              {
                "items": {
                  "type": "string"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Recommended Investigation"
          },
          "containment_options": {
            "anyOf": [
              {
                "items": {
                  "type": "string"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Containment Options"
          },
          "citations": {
            "items": {
              "$ref": "#/components/schemas/Citation"
            },
            "type": "array",
            "title": "Citations"
          },
          "human_review_required": {
            "type": "boolean",
            "const": true,
            "title": "Human Review Required",
            "default": true
          },
          "analysis": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/AIOutput"
              },
              {
                "type": "null"
              }
            ]
          },
          "validation": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Validation"
          },
          "context_snapshot": {
            "additionalProperties": true,
            "type": "object",
            "title": "Context Snapshot"
          },
          "steps": {
            "items": {
              "$ref": "#/components/schemas/StepView"
            },
            "type": "array",
            "title": "Steps"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "run_id",
          "incident_id",
          "tenant_id",
          "status",
          "trusted",
          "validation_status",
          "provider",
          "model",
          "error_code",
          "created_at",
          "started_at",
          "completed_at",
          "data_mode",
          "observed_facts",
          "hypotheses",
          "missing_evidence",
          "recommended_investigation",
          "containment_options",
          "citations",
          "analysis",
          "validation",
          "context_snapshot",
          "steps"
        ],
        "title": "AIRunView"
      },
...
      "AnalyzeIn": {
        "properties": {
          "incident_id": {
            "type": "string",
            "maxLength": 128,
            "minLength": 1,
            "pattern": "^[A-Za-z0-9_.:-]+$",
            "title": "Incident Id"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "incident_id"
        ],
        "title": "AnalyzeIn"
      },
...
      "Citation": {
        "properties": {
          "event_uid": {
            "type": "string",
            "title": "Event Uid"
          },
          "incident_id": {
            "type": "string",
            "title": "Incident Id"
          },
          "source": {
            "type": "string",
            "title": "Source"
          },
          "event_time": {
            "type": "string",
            "format": "date-time",
            "title": "Event Time"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "event_uid",
          "incident_id",
          "source",
          "event_time"
        ],
        "title": "Citation"
      },
...
      "ErrorDetail": {
        "properties": {
          "code": {
            "type": "string",
            "title": "Code"
          },
          "issues": {
            "items": {
              "$ref": "#/components/schemas/ErrorIssue"
            },
            "type": "array",
            "title": "Issues"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "code"
        ],
        "title": "ErrorDetail"
      },
      "ErrorIssue": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Loc"
          },
          "type": {
            "type": "string",
            "title": "Type"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "loc",
          "type"
        ],
        "title": "ErrorIssue"
      },
      "ErrorResponse": {
        "properties": {
          "detail": {
            "$ref": "#/components/schemas/ErrorDetail"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "detail"
        ],
        "title": "ErrorResponse"
      },
...
      "Fact": {
        "properties": {
          "text": {
            "type": "string",
            "maxLength": 2000,
            "minLength": 1,
            "title": "Text"
          },
          "evidence_event_ids": {
            "items": {
              "type": "string",
              "maxLength": 128,
              "minLength": 1,
              "pattern": "^[A-Za-z0-9_.:-]+$"
            },
            "type": "array",
            "maxItems": 20,
            "minItems": 1,
            "title": "Evidence Event Ids"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "text",
          "evidence_event_ids"
        ],
        "title": "Fact"
      },
...
      "Hypothesis": {
        "properties": {
          "text": {
            "type": "string",
            "maxLength": 2000,
            "minLength": 1,
            "title": "Text"
          },
          "limitations": {
            "type": "string",
            "maxLength": 2000,
            "minLength": 1,
            "title": "Limitations"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "text",
          "limitations"
        ],
        "title": "Hypothesis"
      },
...
      "StepView": {
        "properties": {
          "sequence": {
            "type": "integer",
            "title": "Sequence"
          },
          "stage": {
            "type": "string",
            "title": "Stage"
          },
          "status": {
            "type": "string",
            "title": "Status"
          },
          "duration_ms": {
            "type": "integer",
            "title": "Duration Ms"
          },
          "tool_name": {
            "type": "string",
            "title": "Tool Name"
          },
          "output_ref": {
            "type": "string",
            "title": "Output Ref"
          },
          "evidence_ids": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Evidence Ids"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "sequence",
          "stage",
          "status",
          "duration_ms",
          "tool_name",
          "output_ref",
          "evidence_ids"
        ],
        "title": "StepView"
      },
...
    },
    "securitySchemes": {
      "HTTPBearer": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  }
}